feat: shared collection context menu and context menu modal behavior
- Add CollectionContextMenuProvider and wire rename/delete collection menu into sidebar and profile

- Pin collection submenu while collection context menu or any collection modal is open

- Close context menus on scroll in any direction

- Ignore clicks inside modals and other context menus when dismissing

- Darken and blur context menus when a modal is open

- Keep create-collection action from closing the GameContextMenu

- Clean up sidebar focus-visible outline
