+++ src/renderer/src/context/collection-context-menu/index.ts
@@ -1,0 +1,1 @@
+export * from "./collection-context-menu.context";
