+++ src/renderer/src/context/index.ts
@@ -2,3 +2,4 @@ export * from "./game-details/game-details.context";
 export * from "./settings/settings.context";
 export * from "./user-profile/user-profile.context";
 export * from "./cloud-sync/cloud-sync.context";
+export * from "./collection-context-menu";
