+++ src/renderer/src/main.tsx
@@ -13,6 +13,8 @@ import "@fontsource/noto-sans/700.css";
 import "react-loading-skeleton/dist/skeleton.css";
 import "react-tooltip/dist/react-tooltip.css";
 
+import { CollectionContextMenuProvider } from "@renderer/context";
+
 import { App } from "./app";
 
 import { store } from "./store";
@@ -123,47 +125,55 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
     <Provider store={store}>
       <ErrorBoundary>
         <HashRouter>
-          <AchievementNotificationOverlay />
-          <Routes>
-            <Route element={<App />}>
-              <Route path="/" element={<Home />} />
-              <Route path="/catalogue" element={<Catalogue />} />
-              <Route path="/library" element={<Library />} />
-              <Route path="/downloads" element={<Downloads />} />
-              <Route path="/game/:shop/:objectId" element={<GameDetails />} />
-              <Route path="/settings" element={<Settings />} />
-              <Route path="/profile/:userId" element={<Profile />} />
-              <Route path="/achievements" element={<Achievements />} />
-              <Route path="/notifications" element={<Notifications />} />
-            </Route>
+          <CollectionContextMenuProvider>
+            <AchievementNotificationOverlay />
+            <Routes>
+              <Route element={<App />}>
+                <Route path="/" element={<Home />} />
+                <Route path="/catalogue" element={<Catalogue />} />
+                <Route path="/library" element={<Library />} />
+                <Route path="/downloads" element={<Downloads />} />
+                <Route path="/game/:shop/:objectId" element={<GameDetails />} />
+                <Route path="/settings" element={<Settings />} />
+                <Route path="/profile/:userId" element={<Profile />} />
+                <Route path="/achievements" element={<Achievements />} />
+                <Route path="/notifications" element={<Notifications />} />
+              </Route>
 
-            <Route path="/theme-editor" element={<ThemeEditor />} />
-            <Route
-              path="/achievement-notification"
-              element={<AchievementNotification />}
-            />
-            <Route path="/game-launcher" element={<GameLauncher />} />
-            <Route path="/friends-window" element={<FriendsWindow />} />
-            <Route path="/auth-window" element={<AuthWindow />} />
+              <Route path="/theme-editor" element={<ThemeEditor />} />
+              <Route
+                path="/achievement-notification"
+                element={<AchievementNotification />}
+              />
+              <Route path="/game-launcher" element={<GameLauncher />} />
+              <Route path="/friends-window" element={<FriendsWindow />} />
+              <Route path="/auth-window" element={<AuthWindow />} />
 
-            <Route path="/big-picture" element={<BigPictureApp />}>
-              <Route index element={<BigPictureHome />} />
-              <Route path="catalogue" element={<BigPictureCatalogue />} />
-              <Route
-                path="component-lab"
-                element={<BigPictureComponentLab />}
-              />
-              <Route path="downloads" element={<BigPictureDownloads />} />
-              <Route path="settings" element={<BigPictureSettings />} />
-              <Route path="library" element={<BigPictureLibrary />} />
-              <Route path="profile/:userId?" element={<BigPictureProfile />} />
-              <Route path="game/:shop/:objectId" element={<BigPictureGame />} />
-              <Route
-                path="game/:shop/:objectId/achievements"
-                element={<BigPictureGameAchievements />}
-              />
-            </Route>
-          </Routes>
+              <Route path="/big-picture" element={<BigPictureApp />}>
+                <Route index element={<BigPictureHome />} />
+                <Route path="catalogue" element={<BigPictureCatalogue />} />
+                <Route
+                  path="component-lab"
+                  element={<BigPictureComponentLab />}
+                />
+                <Route path="downloads" element={<BigPictureDownloads />} />
+                <Route path="settings" element={<BigPictureSettings />} />
+                <Route path="library" element={<BigPictureLibrary />} />
+                <Route
+                  path="profile/:userId?"
+                  element={<BigPictureProfile />}
+                />
+                <Route
+                  path="game/:shop/:objectId"
+                  element={<BigPictureGame />}
+                />
+                <Route
+                  path="game/:shop/:objectId/achievements"
+                  element={<BigPictureGameAchievements />}
+                />
+              </Route>
+            </Routes>
+          </CollectionContextMenuProvider>
         </HashRouter>
       </ErrorBoundary>
     </Provider>
