+++ src/renderer/src/pages/library/library.tsx
@@ -12,7 +12,6 @@ import {
   useAppDispatch,
   useAppSelector,
   useGameCollections,
-  useToast,
   useUserDetails,
 } from "@renderer/hooks";
 import { setHeaderTitle } from "@renderer/features";
@@ -20,22 +19,13 @@ import {
   HeartIcon,
   TelescopeIcon,
   FileDirectoryIcon,
-  PencilIcon,
-  TrashIcon,
   SyncIcon,
 } from "@primer/octicons-react";
 import { useTranslation } from "react-i18next";
 import { AuthPage } from "@shared";
 import { GameCollection, LibraryGame } from "@types";
-import {
-  Button,
-  ConfirmationModal,
-  ContextMenu,
-  CreateCollectionModal,
-  GameContextMenu,
-  Modal,
-  TextField,
-} from "@renderer/components";
+import { CreateCollectionModal, GameContextMenu } from "@renderer/components";
+import { useCollectionContextMenu } from "@renderer/context";
 import { useSearchParams } from "react-router-dom";
 import { LibraryGameCard } from "./library-game-card";
 import { LibraryGameCardLarge } from "./library-game-card-large";
@@ -87,7 +77,6 @@ const getGameCollectionIds = (game: LibraryGame): string[] => {
 
 export default function Library() {
   const { library, updateLibrary } = useLibrary();
-  const { showSuccessToast, showErrorToast } = useToast();
   const { userDetails } = useUserDetails();
   const {
     collections,
@@ -95,6 +84,7 @@ export default function Library() {
     hasLoaded: hasLoadedCollections,
   } = useGameCollections();
   const [searchParams, setSearchParams] = useSearchParams();
+  const { openCollectionContextMenu } = useCollectionContextMenu();
 
   const [viewMode, setViewMode] = useState<ViewMode>(() => {
     const savedViewMode = localStorage.getItem("library-view-mode");
@@ -113,20 +103,6 @@ export default function Library() {
     visible: boolean;
     position: { x: number; y: number };
   }>({ game: null, visible: false, position: { x: 0, y: 0 } });
-  const [collectionContextMenu, setCollectionContextMenu] = useState<{
-    collection: GameCollection | null;
-    visible: boolean;
-    position: { x: number; y: number };
-  }>({ collection: null, visible: false, position: { x: 0, y: 0 } });
-  const [activeCollection, setActiveCollection] =
-    useState<GameCollection | null>(null);
-  const [showRenameCollectionModal, setShowRenameCollectionModal] =
-    useState(false);
-  const [collectionName, setCollectionName] = useState("");
-  const [isRenamingCollection, setIsRenamingCollection] = useState(false);
-  const [showDeleteCollectionModal, setShowDeleteCollectionModal] =
-    useState(false);
-  const [isDeletingCollection, setIsDeletingCollection] = useState(false);
   const [showCreateCollectionModal, setShowCreateCollectionModal] =
     useState(false);
 
@@ -305,23 +281,6 @@ export default function Library() {
     };
   }, [updateLibrary]);
 
-  const handleOpenCollectionContextMenu = useCallback(
-    (event: React.MouseEvent<HTMLElement>, collection: GameCollection) => {
-      event.preventDefault();
-
-      setCollectionContextMenu({
-        collection,
-        visible: true,
-        position: { x: event.clientX, y: event.clientY },
-      });
-    },
-    []
-  );
-
-  const handleCloseCollectionContextMenu = useCallback(() => {
-    setCollectionContextMenu((prev) => ({ ...prev, visible: false }));
-  }, []);
-
   const handleCreateCollectionButtonClick = useCallback(() => {
     if (!userDetails) {
       window.electron.openAuthWindow(AuthPage.SignIn);
@@ -330,172 +289,6 @@ export default function Library() {
 
     setShowCreateCollectionModal(true);
   }, [userDetails]);
-
-  const resolveCollectionErrorMessage = useCallback(
-    (
-      error: unknown,
-      fallbackKey: "failed_rename_collection" | "failed_delete_collection"
-    ) => {
-      if (!(error instanceof Error)) return t(fallbackKey);
-
-      if (error.message.includes("game/collection-name-already-in-use")) {
-        return t("collection_name_already_in_use", { ns: "sidebar" });
-      }
-
-      if (error.message.includes("game/collection-name-required")) {
-        return t("collection_name_required", { ns: "sidebar" });
-      }
-
-      return t(fallbackKey);
-    },
-    [t]
-  );
-
-  const handleOpenRenameCollectionModal = useCallback(() => {
-    const collection = collectionContextMenu.collection;
-    if (!collection) return;
-
-    setActiveCollection(collection);
-    setCollectionName(collection.name);
-    setShowRenameCollectionModal(true);
-    handleCloseCollectionContextMenu();
-  }, [collectionContextMenu.collection, handleCloseCollectionContextMenu]);
-
-  const handleCloseRenameCollectionModal = useCallback(() => {
-    if (isRenamingCollection) return;
-
-    setShowRenameCollectionModal(false);
-    setCollectionName("");
-    setActiveCollection(null);
-  }, [isRenamingCollection]);
-
-  const handleRenameCollection = useCallback(async () => {
-    if (!activeCollection) return;
-
-    const nextName = collectionName.trim();
-    if (!nextName) {
-      showErrorToast(t("collection_name_required", { ns: "sidebar" }));
-      return;
-    }
-
-    if (nextName === activeCollection.name.trim()) {
-      handleCloseRenameCollectionModal();
-      return;
-    }
-
-    setIsRenamingCollection(true);
-
-    try {
-      await window.electron.hydraApi.put(
-        `/profile/games/collections/${activeCollection.id}`,
-        {
-          data: { name: nextName },
-          needsAuth: true,
-        }
-      );
-
-      await loadCollections();
-      showSuccessToast(t("collection_renamed"));
-      handleCloseRenameCollectionModal();
-    } catch (error) {
-      showErrorToast(
-        resolveCollectionErrorMessage(error, "failed_rename_collection")
-      );
-    } finally {
-      setIsRenamingCollection(false);
-    }
-  }, [
-    activeCollection,
-    collectionName,
-    handleCloseRenameCollectionModal,
-    loadCollections,
-    resolveCollectionErrorMessage,
-    showErrorToast,
-    showSuccessToast,
-    t,
-  ]);
-
-  const handleOpenDeleteCollectionModal = useCallback(() => {
-    const collection = collectionContextMenu.collection;
-    if (!collection) return;
-
-    setActiveCollection(collection);
-    setShowDeleteCollectionModal(true);
-    handleCloseCollectionContextMenu();
-  }, [collectionContextMenu.collection, handleCloseCollectionContextMenu]);
-
-  const handleCloseDeleteCollectionModal = useCallback(() => {
-    if (isDeletingCollection) return;
-
-    setShowDeleteCollectionModal(false);
-    setActiveCollection(null);
-  }, [isDeletingCollection]);
-
-  const handleDeleteCollection = useCallback(async () => {
-    if (!activeCollection) return;
-
-    setIsDeletingCollection(true);
-
-    try {
-      await window.electron.hydraApi.delete(
-        `/profile/games/collections/${activeCollection.id}`,
-        { needsAuth: true }
-      );
-
-      if (selectedCollectionId === activeCollection.id) {
-        handleCollectionSelect(null);
-      }
-
-      await Promise.all([loadCollections(), updateLibrary()]);
-      showSuccessToast(t("collection_deleted"));
-      handleCloseDeleteCollectionModal();
-    } catch (error) {
-      showErrorToast(
-        resolveCollectionErrorMessage(error, "failed_delete_collection")
-      );
-    } finally {
-      setIsDeletingCollection(false);
-    }
-  }, [
-    activeCollection,
-    selectedCollectionId,
-    handleCollectionSelect,
-    loadCollections,
-    updateLibrary,
-    showSuccessToast,
-    t,
-    handleCloseDeleteCollectionModal,
-    showErrorToast,
-    resolveCollectionErrorMessage,
-  ]);
-
-  const collectionContextMenuItems = useMemo(() => {
-    const isCollectionActionBusy = isRenamingCollection || isDeletingCollection;
-
-    return [
-      {
-        id: "rename-collection",
-        label: t("rename_collection"),
-        icon: <PencilIcon size={16} />,
-        onClick: handleOpenRenameCollectionModal,
-        disabled: isCollectionActionBusy,
-      },
-      {
-        id: "delete-collection",
-        label: t("delete_collection"),
-        icon: <TrashIcon size={16} />,
-        onClick: handleOpenDeleteCollectionModal,
-        danger: true,
-        disabled: isCollectionActionBusy,
-      },
-    ];
-  }, [
-    handleOpenDeleteCollectionModal,
-    handleOpenRenameCollectionModal,
-    isDeletingCollection,
-    isRenamingCollection,
-    t,
-  ]);
 
   useEffect(() => {
     if (!selectedCollectionId) return;
@@ -740,7 +533,7 @@ export default function Library() {
                 favoritesCollectionId={FAVORITES_COLLECTION_ID}
                 onSelect={handleCollectionSelect}
                 onCreate={handleCreateCollectionButtonClick}
-                onCollectionContextMenu={handleOpenCollectionContextMenu}
+                onCollectionContextMenu={openCollectionContextMenu}
               />
             </div>
 
@@ -863,71 +656,9 @@ export default function Library() {
           visible={gameContextMenu.visible}
           position={gameContextMenu.position}
           onClose={handleCloseContextMenu}
+          onCollectionContextMenu={openCollectionContextMenu}
         />
       )}
-
-      <ContextMenu
-        items={collectionContextMenuItems}
-        visible={collectionContextMenu.visible}
-        position={collectionContextMenu.position}
-        onClose={handleCloseCollectionContextMenu}
-      />
-
-      <Modal
-        visible={showRenameCollectionModal}
-        title={t("rename_collection")}
-        description={t("rename_collection_description")}
-        onClose={handleCloseRenameCollectionModal}
-      >
-        <div className="library__collection-modal">
-          <TextField
-            label={t("collection_name", { ns: "sidebar" })}
-            placeholder={t("collection_name_placeholder", { ns: "sidebar" })}
-            value={collectionName}
-            onChange={(event) => setCollectionName(event.target.value)}
-            theme="dark"
-            disabled={isRenamingCollection}
-            maxLength={60}
-          />
-
-          <div className="library__collection-modal-actions">
-            <Button
-              type="button"
-              theme="outline"
-              onClick={handleCloseRenameCollectionModal}
-              disabled={isRenamingCollection}
-            >
-              {t("cancel", { ns: "sidebar" })}
-            </Button>
-
-            <Button
-              type="button"
-              theme="primary"
-              onClick={handleRenameCollection}
-              disabled={!collectionName.trim() || isRenamingCollection}
-            >
-              {isRenamingCollection
-                ? t("renaming_collection")
-                : t("rename_collection")}
-            </Button>
-          </div>
-        </div>
-      </Modal>
-
-      <ConfirmationModal
-        visible={showDeleteCollectionModal}
-        title={t("delete_collection_title")}
-        descriptionText={t("delete_collection_description", {
-          collectionName: activeCollection?.name ?? "",
-        })}
-        onClose={handleCloseDeleteCollectionModal}
-        onConfirm={() => {
-          void handleDeleteCollection();
-        }}
-        cancelButtonLabel={t("cancel", { ns: "sidebar" })}
-        confirmButtonLabel={t("delete_collection")}
-        buttonsIsDisabled={isDeletingCollection}
-      />
 
       <CreateCollectionModal
         visible={showCreateCollectionModal}
